requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
git+https://github.com/ozgur/python-firebase
grpcio
google.generativeai
google-genai
google-cloud-storage
Pillow
groq
openai